refactor(app): adjust component structure types, and interfaces
