+++ src/lib/app.pcss
@@ -12,6 +12,7 @@ body {
 }
 
 .card-container {
+  @apply shadow-xl;
   overflow: hidden;
   border-radius: 0.5rem;
   > div {
