+++ src/lib/widgets/CurrentWeather/CurrentWeather.svelte
@@ -5,21 +5,19 @@
   import Barometer from 'svelte-weather/Barometer.svelte';
   import WeatherIcon from './WeatherIcon.svelte';
   import { onMount } from 'svelte';
-  export let weather: {
-    current: number;
-    weatherCode: number[];
-    isDay: boolean;
-  };
+  import type { CurrentWeather } from '$lib/types';
+
+  export let weather: CurrentWeather;
 
   $: ({ current, weatherCode, isDay } = weather);
   $: highlightColor = fahrenheitToColorShade(current);
 </script>
 
-{#if current}
-  <div transition:fade class="card-container relative">
+<div transition:fade class="card-container relative">
+  {#key current}
     <Card size="xl" class="dboard__card">
       <h2>Lansdale</h2>
-      <WeatherIcon {weatherCode} {isDay} color={highlightColor} />
+      <WeatherIcon {weatherCode} {isDay} />
       <h1
         class="mt-auto justify-end text-9xl"
         style={`color: ${highlightColor}`}
@@ -27,5 +25,5 @@
         {Math.round(current)}<span class="text-gray-50">&deg;</span>
       </h1>
     </Card>
-  </div>
-{/if}
+  {/key}
+</div>
